設定エリアの構成パーツをVuetifyに移行
